requirements.txt
requests>=2.26.0
feedparser>=6.0.0
beautifulsoup4>=4.9.3
lxml>=4.9.0
python-dotenv 